+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `bbtw show` command for listing builds in a buildspace ([!185](https://gitlab.archlinux.org/archlinux/buildbtw/-/merge_requests/185))
 
+### repo-updater
+
+-  Fix double color_eyre::install() ([!197](https://gitlab.archlinux.org/archlinux/buildbtw/-/merge_requests/197))
+
 ## [0.0.5] - 2026-05-19
 
 ### backend