+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+## [0.0.6] - 2026-06-04
+
 ### backend
 
 #### Highlights
@@ -98,7 +100,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Just a placeholder for now :)
 
 <!-- next-url -->
-[Unreleased]: https://github.archlinux.org/archlinux/buildbtw/-/compare/v0.0.5...HEAD
+[Unreleased]: https://github.archlinux.org/archlinux/buildbtw/-/compare/v0.0.6...HEAD
+[0.0.6]: https://github.archlinux.org/archlinux/buildbtw/-/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.archlinux.org/archlinux/buildbtw/-/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.archlinux.org/archlinux/buildbtw/-/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.archlinux.org/archlinux/buildbtw/-/compare/v0.0.2...v0.0.3